Make Jenkins stats just show latest build
Previously we would download all the build information from anything in
the past several months and display them all. For the dashboard email we
just want the latest build. We can do that by getting all the build
numbers and use the highest one. This saves a bunch of slow synchronous
requests. We use some simple heuristics to find the latest successful
build, it fails a lot