Add jenkins summary statistics to output
I removed the old jenkins binary script and it's supporting stats module
in favor of modifying the jenkins module directly to return rather than
print data. This is more useful in rendering templates. I avoided using
asyncio because I'd have to rewrite the jenkins client utility that
we're using and I don't want to do that