Add jira summary
I've removed the old jira summary generator from bin because now we will
be putting it in a much nicer format on a static html page. I've moved
some of its relevant functions into the jira module. We don't display a
lot of data yet, but that can grow over time