Add simple cache to render data, add upload script
It's annoying when I'm developing to have to wait to re-gather data
every few minutes when I'm just working on templates. Instead I now have
a cache that lives in ~/.dashi/cache that holds the most recent data.
As-is the cache never invalidates, so you have to remove the file to
invalidate it.

Because I'm too lazy to separate out my commits this also includes the
new upload logic which finds the files that were written as output and
uploads them to S3. This requires some new S3 configuration options in
the config file.