Add config-level repository object
It's a little nicer to deal with, kind of like the User class, as it
gives some common methods for operating on the data like `path`. I also
removed the Repo concept from the git module as it will, over time,
merge into the config.Repository class.

This commit also include some rework to how we validate authors to help
me in rebuilding my config file that I deleted. Instead of listing an
author per commit we just create a set of authors and list them all